Proper Room Spawning
Rooms can spawn in any direction on a 2D grid. This needs to be expanded so rooms know which room comes next and before so later on during gameplay the room knows which door to open after it's complete. This should be done with a doubly linked list where each room has a reference to the previous and upcoming room.