Updated Room Controller and POIs
Each POI now has a POI Data component with a POIData scriptable object attached to it. We use these to tell the parent room what the POI's contain. Additionally, enemies now on death update the room controller's enemy count so it accurately reflects what's in the room. When this and loot reaches 0, the room unlocks and the player can continue.