Fixed player trap trigger stats not saving correctly
The issue was with the data being saved when room was cleared, not when the room was exited. This meant if the player took trap damage after all the enemies were defeated, this statistic was not saved. The data is now saved when room controller invokes room data save which is always before room deleted.